Move middleware files
