+++ packages/next/src/client/components/react-dev-overlay-pages-router/src/internal/helpers/stack-frame.ts
@@ -1,5 +1,5 @@
 import type { StackFrame } from 'next/dist/compiled/stacktrace-parser'
-import type { OriginalStackFrameResponse } from '../../middleware'
+import type { OriginalStackFrameResponse } from '../../middleware/middleware'
 
 export type OriginalStackFrame =
   | {
