+++ packages/next/src/client/components/react-dev-overlay-pages-router/src/middleware/middleware-turbopack.ts
@@ -5,7 +5,7 @@ import type { OriginalStackFrameResponse } from './middleware'
 import fs, { constants as FS } from 'fs/promises'
 import url from 'url'
 import { codeFrameColumns } from 'next/dist/compiled/babel/code-frame'
-import { launchEditor } from './internal/helpers/launchEditor'
+import { launchEditor } from './internal/launchEditor'
 
 interface Project {
   getSourceForAsset(filePath: string): Promise<string | null>
