+++ packages/next/src/server/dev/hot-reloader-webpack.ts
@@ -7,7 +7,7 @@ import type { UrlObject } from 'url'
 import type { RouteDefinition } from '../future/route-definitions/route-definition'
 
 import { webpack, StringXor } from 'next/dist/compiled/webpack/webpack'
-import { getOverlayMiddleware } from '../../client/components/react-dev-overlay-pages-router/src/middleware'
+import { getOverlayMiddleware } from '../../client/components/react-dev-overlay-pages-router/src/middleware/middleware'
 import { WebpackHotMiddleware } from './hot-middleware'
 import { join, relative, isAbsolute, posix } from 'path'
 import {
