+++ packages/next/src/server/lib/router-utils/setup-dev-bundler.ts
@@ -102,11 +102,11 @@ import {
   getErrorSource,
   getSourceById,
   parseStack,
-} from '../../../client/components/react-dev-overlay-pages-router/src/middleware'
+} from '../../../client/components/react-dev-overlay-pages-router/src/middleware/middleware'
 import {
   getOverlayMiddleware,
   createOriginalStackFrame as createOriginalTurboStackFrame,
-} from '../../../client/components/react-dev-overlay-pages-router/src/middleware-turbopack'
+} from '../../../client/components/react-dev-overlay-pages-router/src/middleware/middleware-turbopack'
 import { mkdir, readFile, writeFile } from 'fs/promises'
 import { PageNotFoundError } from '../../../shared/lib/utils'
 import {
