+++ packages/next/src/server/lib/server-ipc/request-utils.ts
@@ -25,7 +25,7 @@ export const deserializeErr = (serializedErr: any) => {
     process.env.NEXT_RUNTIME !== 'edge'
   ) {
     const { decorateServerError } =
-      require('../../../client/components/react-dev-overlay-pages-router/src/middleware') as typeof import('../../../client/components/react-dev-overlay-pages-router/src/middleware')
+      require('../../../client/components/react-dev-overlay-pages-router/src/middleware/middleware') as typeof import('../../../client/components/react-dev-overlay-pages-router/src/middleware/middleware')
     decorateServerError(err, serializedErr.source || 'server')
   }
   return err
