+++ packages/next/src/server/render.tsx
@@ -372,7 +372,7 @@ export function errorToJSON(err: Error) {
   if (process.env.NEXT_RUNTIME !== 'edge') {
     source =
       (
-        require('../client/components/react-dev-overlay-pages-router/src/middleware') as typeof import('../client/components/react-dev-overlay-pages-router/src/middleware')
+        require('../client/components/react-dev-overlay-pages-router/src/middleware/middleware') as typeof import('../client/components/react-dev-overlay-pages-router/src/middleware/middleware')
       ).getErrorSource(err) || 'server'
   }
